requirements.txt
apache-airflow==2.10.4
pyspark==3.5.0
delta-spark==3.0.0
great-expectations==0.17.0
pandas==2.0.0
requests==2.31.0
boto3==1.28.0
# Azure dependencies
azure-identity==1.15.0
azure-storage-file-datalake==12.17.0
azure-storage-blob>=12.23.0
# Additional dependencies for data processing
matplotlib>=3.5.0
seaborn>=0.11.0
python-dotenv>=1.0.0